+++ tests/test_api.py
@@ -520,3 +520,32 @@ async def test_fetch_val_works_ok_with_sa_params(query_executor: AiopgQueryExecu
     res = await query_executor.fetch_val(query, aue=123)
 
     assert res == 123
+
+
+async def test_update_or_insert_many_by_field(repo: TransactionRepo) -> None:
+    await repo.insert(Transaction(price=100, date=dt.date(2020, 1, 1)))
+
+    await repo.update_or_insert_many_by_field(
+        [
+            Transaction(price=200, date=dt.date(2020, 1, 1)),
+            Transaction(price=300, date=dt.date(2021, 2, 2)),
+        ],
+        "date",
+    )
+
+    examples = await repo.get_all()
+    assert len(examples) == 2
+    assert next(i for i in examples if i.date == dt.date(2020, 1, 1)).price == 200
+    assert next(i for i in examples if i.date == dt.date(2021, 2, 2)).price == 300
+
+
+async def test_update_or_insert_first(repo: TransactionRepo) -> None:
+    await repo.insert(Transaction(price=100, date=dt.date(2020, 1, 1)))
+
+    await repo.update_or_insert_first(Transaction(price=200, date=dt.date(2020, 1, 1)), "date")
+    await repo.update_or_insert_first(Transaction(price=300, date=dt.date(2021, 2, 2)), "date")
+
+    examples = await repo.get_all()
+    assert len(examples) == 2
+    assert next(i for i in examples if i.date == dt.date(2020, 1, 1)).price == 200
+    assert next(i for i in examples if i.date == dt.date(2021, 2, 2)).price == 300